+++ Tests/AppTests/PackageController+ShowRouteTests.swift
@@ -427,19 +427,6 @@ class PackageController_ShowRouteTests: AppTestCase {
         XCTAssertEqual(schema.name, "bar")
     }
 
-    func test_show() throws {
-        // setup
-        let pkg = try savePackage(on: app.db, "1")
-        try Repository(package: pkg, name: "package", owner: "owner")
-            .save(on: app.db).wait()
-        try Version(package: pkg, latest: .defaultBranch).save(on: app.db).wait()
-
-        // MUT
-        try app.test(.GET, "/owner/package", afterResponse: { response in
-            XCTAssertEqual(response.status, .ok)
-        })
-    }
-
 }
